ls8: rename .cpu__{constants,masks,operations} <- .*
