+++ ls8/cpu__constants.py
